PHPCodeSniffer support.
Good start, but not sure how to handle cases where phpcs says
the code is clean. The difference is phpcs will exit 0 when code
is clean and non zero otherwise. Don't know how in TidyAll to trap
on that.

A sample of some clean code sniffed by phpcs:

$ tidyall -a
[checked] views/TwigExtensionFilters.class.php
Time: 0 seconds, Memory: 3.00Mb
$ echo $?
0

Change-Id: I15d72fce2223994055dee5aa79c9d96a24e8f996