Catch errors from perltidy even if perltidy -se flag set.
If the perltidy -se flag is set then errors are mixed into the stderr
stream.  If the -nse is set then errors are sent to a separate stream
(errorfile).

This is particularly pernicious because these flags can be set in
other "bundly" flags such as -pbp.

Before this change, if the code only checked the errorfile stream.  If
-se was set then it would miss the errors that were in the stderr stream.

This fix catches both streams and checks both.