emacs mode: clean up temp error-file.
The emacs lisp code creates a temporary file to hold any errors
that tidyall spews.  We should clean it up when we're finished
with it.

My emacs-fu isn't strong enough to deal with cleaning up the file
if something exceptional happens (e.g. what's the elisp equiv. of
a "finally" clause in try/catch/finally?)