use git status --porcelain -z
This was surprisingly obnoxious.

It sounds great, at first, that git uses \0 terminators, but its
records are of variable length, and the record and field separators
are the same -- \0!

So, first you read in the change mode (in format "XY ") and then the
filename ("filename\0").  Then, if the mode was a rename, you read in
a second filename and throw it away.  The first filename is the
renamed-to filename.  (That's backward from the for-humans "x -> y"
format.)

Still, this is less ambiguous.  For example, if there are two ->
arrows in the human-readable "git status," the correct filename to
pick out is indeterminable.  Is it worth adding this complexity to
deal with human boneheadedness?  I leave that decision up to you.