Make the git pre-commit hook only look at files that are actually about to be committed
Previously it would look at everything in the git repo that was modified, even
it wasn't staged to be committed. This made it harder than necessary to do
commits a little bit at a time, as any untidy code in the repo would block
that commit, even if said untidy code was not part of the commit itself.