Add a terrible hack to fix OSX test failures
We need to call ->realpath on the tempdir object we create because OSX uses a
symlink from /tmp to /private/tmp. This causes later path comparisons to fail,
where we compare the ->realpath to a path created via $tempdir->child.

The terrible hack preserves the File::Temp::Dir object in the object returned
by ->realpath.