On Win32 get the long path name for temp dirs
This fixes a number of tests on Windows. Other FS operations were getting
paths with long path names as opposed to short names like "FOOBAR~1". When
Code::TidyAll looked compared the two, it thought that files with long names
were not located under the root dir, which had a short name.