Manipulate PERL5LIB to get t/Basic.t passing again
After introducing the directory change to $ROOT before running commands,
t/Basic.t broke.  The issue seems to be that the command is run in a
different process.  Once the directory change happens, the relative lib
"t/lib" can no longer be found.  In this case, just using FindBin is not
enough to fix the problem.

We could just pass "-I t/lib" as an argument to prove, but that would be
an extra argument to remember.  We could also add it to .proverc, which
means we can get a passing test without the extra argument at the
command line.  This would fix the CI builds.  However, this still does
not fix "dzil test".  This is because "dzil test" runs "make test" and
bypasses prove entirely.  So we need a universal fix which does not rely
on the fact that prove will be used to run the tests.

Messing with PERL5LIB and changing relative paths to absolute is enough
to get these tests in a passing state.

This fix is mostly borrowed from
https://metacpan.org/source/LEONT/Test-Harness-3.39/lib/App/Prove.pm#L586
https://metacpan.org/source/LEONT/Test-Harness-3.39/lib/TAP/Harness.pm#L69
https://metacpan.org/source/LEONT/Test-Harness-3.39/lib/TAP/Parser/SourceHandler/Perl.pm#L162

Thanks to Kent Fredric for working through this with me and basically
just giving me the answers I needed.