Don't tidyall code after dzil changes it and don't test self-tidyall
The problem is that we now have a contributor in the git logs with a UTF-8
character in their name. This gets added to the POD in `lib/Code/Tidyall.pm`
by the Git::Contributors dzil plugin. Then we tidy that file and the UTF-8
gets mangled. See GH #84 for related discussion, as well as
https://blog.urth.org/2020/04/25/the-real-dirt-on-tidyall/.

All of this leads to tidyall failing its self-test for tidyness on post-dzil
build, which in turn blocks making a release.

I tried a bunch of things to fix this, but fundamentally I think it requires
tidyall to replace all use of `(slurp|spew)_raw` with `(slurp|spew)_utf8`. And
that's a _huge_ change that would probably break something for someone.

So I'm making this gross change so I can get a release with some small bug
fixes out the door.