specifically override default branch name
Setting GIT_CONFIG_GLOBAL to prevent loading of ~/.gitconfig isn't
suppoted by all versions of git; on those one may still get test failures
if the user has specified a init.defaultBranch configuration value which
is different from the test expects.

Explicitly pass it on the 'git' command line to ensure that it is
indepndent of the user's configuration.

use the default branch name of 'default-branch' to make it obvious.